pyproject.toml
[project]
name = "localred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "RoCry", email = "crysheen@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "playwright>=1.50.0",
    "pydantic>=2.10.6",
]

[project.scripts]
localred = "localred:main"

[project.optional-dependencies]
example = [
    "fire>=0.7.0",
    "smolllm",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
extend-select = ["I"]

[tool.uv.sources]
smolllm = { path = "../smolllm" }
